refactor(TemplateUploadPage): TemplateEdit 및 useTemplateUpload 삭제
