--- frontend/src/pages/TemplateUploadPage/hooks/index.ts
@@ -1,1 +1,0 @@
-export { useTemplateUpload } from './useTemplateUpload';
